+++ terminaler-gui/src/overlay/mod.rs
@@ -11,6 +11,7 @@ pub mod confirm_close_pane;
 pub mod copy;
 pub mod debug;
 pub mod launcher;
+pub mod new_agent;
 pub mod prompt;
 pub mod quickselect;
 pub mod selector;